requirements.txt
apiview==1.3.25
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-import-export==3.3.8
djangorestframework==3.15.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
sqlparse==0.5.0
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
